Add support for GSF runs in different directories
This is useful for longer jobs where you end up wanting
to run multiple calculations in parallel. In this case it
is a good idea to keep each calculation in a different
directory (some parallel file systems work better this way,
and not all codes are good at not using the same name for
output files for different jobs).

We already support this for setup - this adds support for
data extraction, giving read_gsf.py the --indir argument.